+++ packages/media-parser/src/boxes/riff/expect-riff-box.ts
@@ -1,51 +1,19 @@
 import type {BufferIterator} from '../../buffer-iterator';
-import {getTracks} from '../../get-tracks';
 import type {Options, ParseMediaFields} from '../../options';
 import {
 	registerTrack,
 	registerVideoTrackWhenProfileIsAvailable,
 } from '../../register-track';
 import type {ParserState} from '../../state/parser-state';
-import {
-	makeAviAudioTrack,
-	makeAviVideoTrack,
-	TO_BE_OVERRIDDEN_LATER,
-} from './get-tracks-from-avi';
+import {makeAviAudioTrack, makeAviVideoTrack} from './get-tracks-from-avi';
 import {isMoviAtom} from './is-movi';
-import {parseMovi} from './parse-movi';
 import {parseRiffBox} from './parse-riff-box';
+import {parseVideoSection} from './parse-video-section';
 import type {RiffBox} from './riff-box';
 
-export type RiffResult =
-	| {
-			type: 'incomplete';
-	  }
-	| {
-			type: 'complete';
-			box: RiffBox | null;
-			skipTo: number | null;
-	  };
-
-const parseVideoSection = ({
-	state,
-	iterator,
-}: {
-	state: ParserState;
-	iterator: BufferIterator;
-}) => {
-	const videoSection = state.videoSection.getVideoSection();
-
-	const movi = parseMovi({
-		iterator,
-		maxOffset: videoSection.start + videoSection.size,
-		state,
-	});
-	const tracks = getTracks(state.structure.getStructure(), state);
-	if (!tracks.videoTracks.some((t) => t.codec === TO_BE_OVERRIDDEN_LATER)) {
-		state.callbacks.tracks.setIsDone();
-	}
-
-	return movi;
+export type RiffResult = {
+	box: RiffBox | null;
+	skipTo: number | null;
 };
 
 export const expectRiffBox = async ({
@@ -60,14 +28,9 @@ export const expectRiffBox = async ({
 	// Need at least 16 bytes to read LIST,size,movi,size
 	if (iterator.bytesRemaining() < 16) {
 		return {
-			type: 'incomplete',
+			box: null,
+			skipTo: null,
 		};
-	}
-
-	const isInsideVideoSection =
-		state.videoSection.isInVideoSectionState(iterator);
-	if (isInsideVideoSection === 'in-section') {
-		return parseVideoSection({state, iterator});
 	}
 
 	const checkpoint = iterator.startCheckpoint();
@@ -88,7 +51,8 @@ export const expectRiffBox = async ({
 	if (iterator.bytesRemaining() < ckSize) {
 		checkpoint.returnToCheckpoint();
 		return {
-			type: 'incomplete',
+			box: null,
+			skipTo: null,
 		};
 	}
 
@@ -130,7 +94,6 @@ export const expectRiffBox = async ({
 	}
 
 	return {
-		type: 'complete',
 		box,
 		skipTo: null,
 	};
