+++ packages/media-parser/src/boxes/riff/riff-box.ts
@@ -86,10 +86,6 @@ export type RiffHeader = {
 	fileType: string;
 };
 
-export type MoviBox = {
-	type: 'movi-box';
-};
-
 export type IsftBox = {
 	type: 'isft-box';
 	software: string;
@@ -104,5 +100,4 @@ export type RiffBox =
 	| StrhBox
 	| StrfBoxVideo
 	| StrfBoxAudio
-	| MoviBox
 	| IsftBox;
