+++ packages/media-parser/src/parse-media.ts
@@ -122,8 +122,12 @@ export const parseMedia: ParseMedia = async function <
 		}
 
 		if (iterator.counter.getOffset() === contentLength) {
-			Log.verbose(logLevel, 'Reached end of file');
-			return true;
+			Log.verbose(logLevel, 'Reached end of file', contentLength);
+			// TODO: Make it possible for ISO base media too
+			return (
+				state.structure.getStructureOrNull()?.type !== 'iso-base-media' &&
+				state.structure.getStructureOrNull()?.type !== 'transport-stream'
+			);
 		}
 
 		return false;
